Added warning for docker desktop settings
