+++ ui/src/components/Views/Configs/index.tsx
@@ -1,3 +1,3 @@
 export * from './ConfigPage';
 export * from './UpsertConfig';
-export * from './MountPointForm';
+export * from './SettingsForm';
