+++ ui/src/constants/common.ts
@@ -1,4 +1,3 @@
 export const DEFAULT_CONFIGURATION_ID='00000000-0000-0000-0000-000000000000';
 export const CORS_ALLOW_DEFAULT='http://localhost:3000';
 export const LATEST_IMAGE = 'localstack/localstack:latest';
-export const ERROR_USER = 'DEFAULT_ERROR_USER_LOCATION_FAILED'; // user by default cannot have more than 30 char so this prevents accidentally selecting a real user
